Switch credits from int to long

+++ app/src/main/java/fr/corenting/edcompanion/models/Credits.java
@@ -1,6 +1,6 @@
 package fr.corenting.edcompanion.models;
 
 public class Credits extends BaseModel {
-    public int Balance;
-    public int Loan;
+    public long Balance;
+    public long Loan;
 }
